backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
flask-marshmallow==1.3.0
flask-migrate==4.1.0
flask-sqlalchemy==3.1.1
marshmallow-sqlalchemy==1.4.2
psycopg2-binary==2.9.11
python-dotenv==1.1.1
